Remove Markdown sanitization; sanitize HTML output only
Trying to sanitize Markdown (using `sanitize-markdown` or anything else)
in such a way that it produces safe HTML, without mangling perfectly
safe Markdown documents, is difficult. #182 is just one example; the
hack to re-instate `>` characters for blockquotes is another.

Moreover, we were already sanitizing the HTML output, albeit with the
`xss` library. All it takes is allowing the HTML tags that are produced
by Markdown itself.

Fixes #182.