Remove redirect if POST happened without valid CSRF token
- For posts happening through an old-fashioned browser form submit, the
  redirect makes it look like the post succeeded but forgot or ignored
  all our changes.

- For posts happening through XHR, the redirect makes it look like the
  post succeeded even though it failed.

The best solution would be to render the same page again, but with an
error message stuck into it somewhere. But that requires tedious
per-page changes.

Bonus change in this commit: if an error contain a status code, use that
status code rather than assuming 500. (Makes CSRF errors give 403
instead.)